feat!: create transaction model to replace transfer, deposit, and withdrawal, and move to separate files and folders
