--- server/src/models/Deposit.js
@@ -1,19 +1,0 @@
-import mongoose, { Schema } from "mongoose";
-
-const depositSchema = new mongoose.Schema(
-  {
-    accountId: {
-      type: "string",
-      required: true,
-    },
-    amount: {
-      type: Schema.Types.Decimal128,
-      required: true,
-    },
-  },
-  { timestamps: true }
-);
-
-const Deposit = mongoose.model("Deposit", depositSchema);
-
-export default Deposit;
