+++ server/src/routes/index.js
@@ -2,9 +2,9 @@ import express from "express";
 import RegisterRouter from "./auth/register.js";
 import LoginRouter from "./auth/login.js";
 import CreateNewAccountRouter from "./account/newAccount.js";
-import TransferFundsRouter from "./account/transferFunds.js";
-import DepositRouter from "./account/depositFunds.js";
-import WithdrawRouter from "./account/withdrawFunds.js";
+import TransferFundsRouter from "./transaction/transferFunds.js";
+import DepositRouter from "./transaction/depositFunds.js";
+import WithdrawRouter from "./transaction/withdrawFunds.js";
 
 const router = express.Router();
